EVEREST-1471 cleanup
